Task 7 condition initial

+++ 0x1E-search_algorithms/100-jump.c
@@ -30,7 +30,7 @@ int jump_search(int *array, size_t size, int value)
 	}
 	printf("Value checked array[%ld] = [%ld]\n", prev, array[prev]);
 	if (array[prev] == value)
-		return (prev);
+		return ((int)prev);
 
 	return (-1);
 }
